Added storing with serelizer
